refactor(ersatz): use @formidable-webview/skeletton extracted logic
BREAKING CHANGE: The type of the default exported component has the same
props, but it is not a class anymore. Instead, exotic forwarded ref
component. That shouldn't change anything for a vast majority of users.
If you are using @formidable-webview/ersatz-testing though, you will
need to upgrade that one too.