fix(ersatz): upgrade jsdom dependency to 16.4.0
fix #1
The dependency version was set to a branch, which caused npm to fail
because the authors of jsdom use a non-standard yarn protocol in one of
their devDependencies. The issue does not seems to happen on versions
from npm. The reason why the branch was used is not relevant anymore.